requirements.txt
django
django_select2
django-bootstrap-v5
django_redis
Pillow
gunicorn
flake8
python-dateutil
psycopg2-binary